Fixed FV1.21.3 query bug

+++ init/framework/1.21.3.php
@@ -3,6 +3,6 @@
  * Fix servers table, createdby may be NULL
  */
 sql_query('ALTER TABLE `users`
-           CHANGE COLUMN `latitude`  `latitude`  DOUBLE(10, 14) NULL DEFAULT NULL,
-           CHANGE COLUMN `longitude` `longitude` DOUBLE(10, 14) NULL DEFAULT NULL');
+           CHANGE COLUMN `latitude`  `latitude`  DOUBLE(14, 14) NULL DEFAULT NULL,
+           CHANGE COLUMN `longitude` `longitude` DOUBLE(14, 14) NULL DEFAULT NULL');
 ?>
